Revisi 1 InventoryService

+++ futureprogram/src/main/java/com/oims/futureprogram/service/impl/InventoryServiceImpl.java
@@ -1,20 +1,16 @@
-package com.oims.futureprogram.service;
+package com.oims.futureprogram.service.impl;
 
 import com.oims.futureprogram.exception.ResourceNotFoundException;
 import com.oims.futureprogram.model.Inventory;
 import com.oims.futureprogram.repository.InventoryRepository;
+import com.oims.futureprogram.service.InventoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.stereotype.Service;
 
-import javax.validation.Valid;
-
-@RestController
-public class InventoryServiceImplement implements  InventoryService{
+@Service
+public class InventoryServiceImpl implements InventoryService {
 
     @Autowired
     private InventoryRepository inventoryRepository;
@@ -23,15 +19,15 @@ public class InventoryServiceImplement implements  InventoryService{
         return inventoryRepository.findAll(pageable);
     }
 
-    public Inventory getOneInventory(@PathVariable Long inventoryId) {
+    public Inventory getOneInventory(Long inventoryId) {
         return inventoryRepository.findById(inventoryId).orElseThrow(() -> new ResourceNotFoundException("Inventory not found with Id " + inventoryId));
     }
 
-    public Inventory createInventory(@Valid @RequestBody Inventory inventory) {
+    public Inventory createInventory(Inventory inventory) {
         return inventoryRepository.save(inventory);
     }
 
-    public Inventory updateInventory(@PathVariable Long inventoryId, @Valid @RequestBody Inventory inventoryRequest) {
+    public Inventory updateInventory(Long inventoryId, Inventory inventoryRequest) {
         return inventoryRepository.findById(inventoryId).map(inventory -> {
             inventory.setNama(inventoryRequest.getNama());
             inventory.setHarga(inventoryRequest.getHarga());
@@ -41,10 +37,11 @@ public class InventoryServiceImplement implements  InventoryService{
         }).orElseThrow(() -> new ResourceNotFoundException("Inventory not found with Id " + inventoryId));
     }
 
-    public ResponseEntity<?> deleteInventory(@PathVariable Long inventoryId) {
-        return inventoryRepository.findById(inventoryId).map(inventory -> {
+    @Override
+    public void deleteInventory(Long inventoryId) {
+        Inventory inventory = getOneInventory(inventoryId);
+        if (inventory != null) {
             inventoryRepository.delete(inventory);
-            return ResponseEntity.ok().build();
-        }).orElseThrow(() -> new ResourceNotFoundException("Inventory not found with Id " + inventoryId));
+        }
     }
 }
